prepull

+++ src/main/java/com/zipcodewilmington/scientificcalculator/Scientific.java
@@ -31,7 +31,7 @@ public class Scientific {
     }
 
 
-    public String getCon (){
+    public String getCon(){
         return con;
     }
 
@@ -79,19 +79,19 @@ public class Scientific {
         return this.currentVal;
     }
 
-    public double getSine(int num4){
+    public double getSine(double num4){
         this.currentVal = sin(num4);
         return this.currentVal;
 
     }
 
-    public double getCosine(int num5){
+    public double getCosine(double num5){
         this.currentVal = cos(num5);
         return this.currentVal;
 
     }
 
-    public double getTangent(int num6){
+    public double getTangent(double num6){
         this.currentVal = tan(num6);
         return this.currentVal;
